+++ js/utils/calc-points.js
@@ -1,9 +1,16 @@
 const BASE_VALUE = {
+  quickAnswerTime: 30,
   correctAnswer: 1,
   quickCorrectAnswer: 2,
-  falseAnswer: -2,
+  falseAnswer: 2,
   baseNotes: 3,
   fullGame: 10
+};
+
+export const quickAnswers = (array) => {
+  return array.filter((it) => {
+    return it.time <= BASE_VALUE.quickAnswerTime;
+  }).length;
 };
 
 const calcPoints = (arr, notes) => {
@@ -13,12 +20,12 @@ const calcPoints = (arr, notes) => {
   }
 
   arr.forEach((it) => {
-    if (it.answer === true && it.time < 30) {
+    if (it.success === true && it.time < BASE_VALUE.quickAnswerTime) {
       points += BASE_VALUE.quickCorrectAnswer;
     }
-    if (it.answer === true && it.time > 30) {
+    if (it.success === true && it.time > BASE_VALUE.quickAnswerTime) {
       points += BASE_VALUE.correctAnswer;
-    } else if (it.answer === false) {
+    } else if (it.success === false) {
       points -= BASE_VALUE.falseAnswer;
     }
